Actually, solve this in a lighter-weight way
The templating is very superficial
