+++ src/libstore/sqlite.hh
@@ -102,11 +102,21 @@ struct SQLiteError : Error
     int errNo, extendedErrNo;
 
     template<typename... Args>
-    [[noreturn]] static void throw_(sqlite3 * db, const std::string & fs, const Args & ... args);
+    [[noreturn]] static void throw_(sqlite3 * db, const std::string & fs, const Args & ... args) {
+        throw_(db, hintfmt(fs, args...));
+    }
 
 protected:
+
+    SQLiteError(const char *path, int errNo, int extendedErrNo, hintformat && hf);
+
     template<typename... Args>
-    SQLiteError(const char *path, int errNo, int extendedErrNo, const Args & ... args);
+    SQLiteError(const char *path, int errNo, int extendedErrNo, const std::string & fs, const Args & ... args)
+      : SQLiteError(path, errNo, extendedErrNo, hintfmt(fs, args...))
+    { }
+
+    [[noreturn]] static void throw_(sqlite3 * db, hintformat && hf);
+
 };
 
 MakeError(SQLiteBusy, SQLiteError);
